Feat: 코스 상세 조회 구현
