+++ src/main/java/com/example/ondongnae/backend/course/dto/CourseDetailDto.java
@@ -1,0 +1,16 @@
+package com.example.ondongnae.backend.course.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@Builder
+public class CourseDetailDto {
+
+    private String title;
+    private String description;
+    private List<RecommendedCourseStoreDto> recommendedCourseStores;
+
+}
